pyproject.toml
[project]
name = "repl-nix-automatedmediaflow"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "trafilatura>=1.12.2",
    "openai>=1.51.2",
    "instagram-private-api>=1.6.0.0",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.149.0",
    "facebook-sdk>=3.1.0",
    "tweepy>=4.14.0",
    "flask-socketio>=5.4.1",
    "google-auth>=2.35.0",
    "stripe>=11.1.0",
    "flask-mail>=0.10.0",
    "textblob>=0.18.0.post0",
]
